archive/legacy/deploy_debug_20250722_015612/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.2
email-validator==2.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML Libraries for conversational agents (auto-resolved versions)
langchain
langgraph
langchain-google-genai
langchain-openai
langchain-core
openai
google-generativeai

# Utilities
requests==2.31.0
httpx==0.25.2

# Azure Application Insights
applicationinsights

# Additional dependencies for conversational flow
typing-extensions>=4.9.0
json5==0.9.14
